refactor(daffio): migrate why-pwa to standalone
https://github.com/graycoreio/daffodil/issues/3971
