--- apps/daffio/src/app/content/why-pwa/view/why-pwa-view.component.ts
@@ -1,12 +1,0 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-} from '@angular/core';
-
-@Component({
-  selector: 'daffio-why-pwa-view',
-  templateUrl: './why-pwa-view.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: false,
-})
-export class DaffioWhyPwaViewComponent {}
