+++ apps/daffio/src/app/content/why-pwa/why-pwa-view.component.spec.ts
@@ -16,8 +16,6 @@ describe('DaffioWhyPwaViewComponent', () => {
     TestBed.configureTestingModule({
       imports: [
         RouterTestingModule,
-      ],
-      declarations: [
         DaffioWhyPwaViewComponent,
       ],
       schemas: [
