--- apps/daffio/src/app/content/why-pwa/why-pwa.module.ts
@@ -1,22 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-
-import { DaffioWhyPwaViewModule } from './view/why-pwa-view.modules';
-import { DaffioWhyPwaRoutingModule } from './why-pwa-routing.module';
-import { TemplateModule } from '../../core/template/template.module';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    RouterModule,
-
-    TemplateModule,
-
-    DaffioWhyPwaRoutingModule,
-    DaffioWhyPwaViewModule,
-  ],
-})
-export class DaffioWhyPwaModule { }
